refactor: sort army builder steps
